refactor: reorganize Go entrypoints
