+++ system/cmd/lambda/error_log_notify_discord/main_test.go
